WIP: media: rockchip: rkisp1: add support for rkisp2.1
The Rockchip RK356x SoCs feature version v2.1 of the Rockchip ISP.

NB: It has not been discussed yet whether the rkisp1 driver should
support rkisp versions > 1. But in order to evaluate that, bring up
the ISP on RK356x with a few crude patches. Please do not run these
patches on rkisp1 HW such as the RK3599.

Signed-off-by: Michael Riesch <michael.riesch@wolfvision.net>
Co-developed-by: Gerald Loacker <gerald.loacker@wolfvision.net>
Signed-off-by: Gerald Loacker <gerald.loacker@wolfvision.net>